pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ragPDF"
version = "0.0.1"
requires-python = ">= 3.12.0"
authors = [
  {name="Dr. Niels Ehlen", email="author@example.com"}
]
description = "A package for analyzing PDFs with GenAI"
readme = "README.md"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

dependencies = [
    "pypdf", "pillow", "pymupdf",
]

[project.optional-dependencies]
linting = [
  'pylint'
]
formatting = [
  'black[d]'
]
dev = ['my_application_name[linting, formatting]']